+++ DV1572/DV1572/trolls_inn/Bobby test/Source2(bobby test).cpp
@@ -52,7 +52,7 @@ int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPWSTR lpCmdLi
 	
 
 	Window wnd(hInstance);
-	wnd.Init(gameSettings.width, gameSettings.height, "Trolls_inn", gameSettings.fullscreen, working);
+	wnd.Init(static_cast<int>(gameSettings.width), static_cast<int>(gameSettings.height), "Trolls_inn", gameSettings.fullscreen, working);
 	using namespace std::chrono;
 	auto time = steady_clock::now();
 	auto timer = steady_clock::now();
